working on get method in users table

+++ src/main/java/org/test/sms/server/service/general/UserGroupServiceImpl.java
@@ -4,9 +4,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.test.sms.common.entity.general.UserGroup;
+import org.test.sms.common.enums.general.ErrorCode;
 import org.test.sms.common.exception.AppException;
 import org.test.sms.common.filter.AbstractFilter;
 import org.test.sms.common.service.general.UserGroupService;
+import org.test.sms.common.service.general.UserService;
 import org.test.sms.server.dao.interfaces.general.UserGroupDao;
 
 import java.util.List;
@@ -18,9 +20,12 @@ public class UserGroupServiceImpl implements UserGroupService {
 
     private UserGroupDao dao;
 
+    private UserService userService;
+
     @Autowired
-    public UserGroupServiceImpl(UserGroupDao dao) {
+    public UserGroupServiceImpl(UserGroupDao dao, UserService userService) {
         this.dao = dao;
+        this.userService = userService;
     }
 
     @Override
@@ -35,6 +40,10 @@ public class UserGroupServiceImpl implements UserGroupService {
 
     @Override
     public void delete(long id) throws AppException {
+        if (userService.exists(id)) {
+            throw new AppException(ErrorCode.USER_GROUP_HAS_USERS);
+        }
+
         dao.delete(id);
     }
